Add window dimensions to Constants
The window dimensions in main.cpp and GLWindow.hpp have been replaced with constants. A new file, Constants.hpp, has been created and added to CMakeLists.txt for managing these constants. This change helps to centralize all the constant values used across different components in one file, thereby making it easier to find and alter these numbers in the future when we need to change the window size or introduce new constants.