Integrate vcpkg setup into CMake workflow
This commit incorporates the setup of vcpkg into the CMake workflow by utilizing the 'run-vcpkg' action. This addition ensures that the right vcpkg files are fetched before the configuration of CMake commences. Furthermore, it adds steps to list the working directory before and after the build, providing visibility into the workspace structure and contents at key phases of the workflow.