Switch declaration of window to pointer in GLGame
The GLGame class now uses a pointer declaration for the variable 'window', instead of a reference. This was done to facilitate correct use of the window object in GLGame functions. The ordering of initializations for camera, scene and entity in 'Runner.hpp' has also been adjusted for better organization and clarity.